Add support for the new Rack specification
This adds native WebSocket support to ActiveCable, based on the [Rack PR #1272](https://github.com/rack/rack/pull/1272).

Tested using iodine version 0.6.4.

IMHO, this is the best **compatible** approach, that still accommodates `hijack` support.

This approach should improve server-client performance. However, channel and broadcasting performance is likely to experience the same bottlenecks that aren't addressed by this change.

Please note:

This approach moves the responsibility for the WebSocket and network logic to the Server, freeing the application and framework from protocol and network concerns.

IMHO, it would make sense if the road-map were updated so `hijack` support would be dropped after the default server (currently Puma) implements the new Rack specification.

This would allow Rails to reduce it's dependencies and remove the `websocket-driver` and the `nio4r` gems from the dependency list.

This would also improve the separation of concerns, where the framework could be freed from IO or protocol related concerns.